shortcuts: start on new shortcuts engine
This starts laying the groundwork for the basic plumbing. In general, we
have an IdeShortcutManager singleton for the process. This tracks the
current keytheme and provides an easy entry point for shortcut consuming
machinery.

Widgets gain the concept of an IdeShortcutController, who processes
incoming events and dispatches them to the current IdeShortcutContext
for the controller. By using the "set-context" signal on the controller,
widgets can change their context before processing the next signal
(however, the signal chain will continue from the original context).

The IdeWorkbench dispatches key-press-events to the shortcut manager who
returns TRUE if they were handled. This could happen because of the current
focus chain, or another widget in the hierarchy whose controller was
registered with the toplevel.

To chain multiple shortcut handlers together, simply call the _add_*()
variant on IdeShortcutContext with your additional operation. All
registered entries for that shortcut will be called in the order they were
added to the context. So to call an action, then a signal, do:

  ide_shortcut_context_add_action(context, "<primary>c", …)
  ide_shortcut_context_add_signal(context, "<primary>c", …)

Some things to think about still:

 - We need the concept of a "keybinding slot" so that it is easy to
   provide UI to override or change keybindings. That means, to some
   degree, providing title/subtitle information when registering
   accelerators. We also need to find a way to hook that up into the
   shortcuts window (so we need section/group as well).

 - We need code to integrate into the EggMenuManager and update the
   accelerator information so that menus are displayed with proper
   accel labels.

 - We need a new IdeShortcutLabel that can be used in tooltips and other
   places to show keybindings similar to the CSS I made for the
   shortcuts window.

 - Now that we will have acclerators registered at the toplevel, we can
   theoretically do a "highlight mode" that shows accelerators and what
   they affect in the UI. But doing this cleanly becomes a fun layout
   challenge more than anything.