Run the modified application on a different session bus
The modified application tries to run on the same session bus
as the one we are forking from, and if the application is still
running this can lead to side effects (e.g. creating another window
of the same app). This workaround uses "dbus-run-session" to start
a session bus instance of dbus-daemon and run our application
in that session.

This has the downside that the application does not have
access to anything on the session bus after that.