build: remove libidemm
Nothing is using libidemm and we won't be needing it for the things we
have upcoming. Babysitting gtkmm/glibmm and it's lack of support for
proper out-of-tree builds is getting annoying, so until something decides
to implement stuff in C++, I'm just going to drop this.

Having all of this in a single commit is convenience in case we want to
revert it at some point in the future.

Additionally, it's reliance on automake is holding back our ability to
move to Meson/Ninja in the future.